pyproject.toml
[project]
name = "macrozero"
version = "0.1.0"
description = "AI code reviews that actually work."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-adk>=1.13.0",
    "litellm>=1.77.0",
    "lmstudio>=1.5.0",
    "pandas>=2.3.2",
    "pyjwt[crypto]>=2.10.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "pytidb[models]>=0.0.13",
    "requests>=2.32.5",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
    "psycopg2-binary>=2.9.11",
    "redis>=7.1.0",
    "rq>=2.6.1",
]
